Break apart two commands within code block
The code block gave the impression that these two commands would be run from a single command line execution.  However, the first command will not return, and even if we were using '&&', the second command won't run.
By breaking the code-block, placing the '&' on a different line, and then starting a new code-block, we add clarity.

BTW. This was the best redis-sentinel "get started" page I found.  Thank you!